+++ lib/base_style_sheet.dart
@@ -31,7 +31,6 @@ export 'package:base_style_sheet/src/presentation/widgets/containers/local_theme
 export 'package:base_style_sheet/src/presentation/widgets/custom_app_bar.dart';
 export 'package:base_style_sheet/src/presentation/widgets/custom_checkbox_tile.dart';
 export 'package:base_style_sheet/src/presentation/widgets/custom_dialog.dart';
-export 'package:base_style_sheet/src/presentation/widgets/custom_dropdown/custom_dropdown.dart';
 export 'package:base_style_sheet/src/presentation/widgets/custom_item_tile.dart';
 export 'package:base_style_sheet/src/presentation/widgets/custom_loading.dart';
 export 'package:base_style_sheet/src/presentation/widgets/custom_radio_tile.dart';
@@ -43,6 +42,7 @@ export 'package:base_style_sheet/src/presentation/widgets/custom_web_view.dart';
 export 'package:base_style_sheet/src/presentation/widgets/custom_wrap.dart';
 export 'package:base_style_sheet/src/presentation/widgets/dividers/custom_divider.dart';
 export 'package:base_style_sheet/src/presentation/widgets/dividers/dashed_divider.dart';
+export 'package:base_style_sheet/src/presentation/widgets/dropdown/custom_dropdown.dart';
 export 'package:base_style_sheet/src/presentation/widgets/empties/list_empty.dart';
 export 'package:base_style_sheet/src/presentation/widgets/errors/custom_request_error.dart';
 export 'package:base_style_sheet/src/presentation/widgets/expansions/custom_expansion.dart';